Streaming: rename all files and references to proper casing
