+++ libraries/botframework-streaming/src/disassemblers/httpContentStreamDisassembler.ts
@@ -5,12 +5,12 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-import { HttpContentStream } from '../HttpContentStream';
-import { PayloadSender } from '../PayloadTransport/PayloadSender';
-import { SubscribableStream } from '../SubscribableStream';
-import { PayloadTypes } from '../Payloads/PayloadTypes';
-import { PayloadDisassembler } from './PayloadDisassembler';
-import { IStreamWrapper } from '../Interfaces/IStreamWrapper';
+import { HttpContentStream } from '../httpContentStream';
+import { PayloadSender } from '../payloadTransport/payloadSender';
+import { SubscribableStream } from '../subscribableStream';
+import { PayloadTypes } from '../payloads/payloadTypes';
+import { PayloadDisassembler } from './payloadDisassembler';
+import { IStreamWrapper } from '../interfaces/IStreamWrapper';
 
 export class HttpContentStreamDisassembler extends PayloadDisassembler {
     public readonly contentStream: HttpContentStream;
