+++ libraries/botframework-streaming/src/disassemblers/payloadDisassembler.ts
@@ -5,11 +5,11 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-import { IHeader } from '../Interfaces/IHeader';
-import { PayloadTypes } from '../Payloads/PayloadTypes';
-import { PayloadSender } from '../PayloadTransport/PayloadSender';
-import { SubscribableStream } from '../SubscribableStream';
-import { IStreamWrapper } from '../Interfaces/IStreamWrapper';
+import { IHeader } from '../interfaces/IHeader';
+import { PayloadTypes } from '../payloads/payloadTypes';
+import { PayloadSender } from '../payloadTransport/payloadSender';
+import { SubscribableStream } from '../subscribableStream';
+import { IStreamWrapper } from '../interfaces/IStreamWrapper';
 
 export abstract class PayloadDisassembler {
     public abstract payloadType: PayloadTypes;
