+++ libraries/botframework-streaming/src/index.ts
@@ -5,12 +5,12 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-export { ContentStream } from './ContentStream';
-export { HttpContent } from './HttpContentStream';
+export { ContentStream } from './contentStream';
+export { HttpContent } from './httpContentStream';
 export { IStreamingTransportServer, IStreamingTransportClient, ISocket, IReceiveRequest, IReceiveResponse } from './Interfaces'
-export { NamedPipeClient, NamedPipeServer } from './NamedPipe';
-export { RequestHandler } from './RequestHandler';
-export { StreamingRequest } from './StreamingRequest';
-export { StreamingResponse } from './StreamingResponse';
-export { SubscribableStream } from './SubscribableStream';
-export { BrowserWebSocket, NodeWebSocket, WebSocketClient, WebSocketServer } from './WebSocket';
+export { NamedPipeClient, NamedPipeServer } from './namedPipe';
+export { RequestHandler } from './requestHandler';
+export { StreamingRequest } from './streamingRequest';
+export { StreamingResponse } from './streamingResponse';
+export { SubscribableStream } from './subscribableStream';
+export { BrowserWebSocket, NodeWebSocket, WebSocketClient, WebSocketServer } from './webSocket';
