+++ libraries/botframework-streaming/src/interfaces/IStreamWrapper.ts
@@ -5,7 +5,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-import { SubscribableStream } from '../SubscribableStream';
+import { SubscribableStream } from '../subscribableStream';
 
 export interface IStreamWrapper {
     stream: SubscribableStream;
