+++ libraries/botframework-streaming/src/namedPipe/namedPipeClient.ts
@@ -6,16 +6,16 @@
  * Licensed under the MIT License.
  */
 import { connect } from 'net';
-import { ProtocolAdapter } from '../ProtocolAdapter';
-import { RequestHandler } from '../RequestHandler';
-import { StreamingRequest } from '../StreamingRequest';
-import { RequestManager } from '../Payloads';
+import { ProtocolAdapter } from '../protocolAdapter';
+import { RequestHandler } from '../requestHandler';
+import { StreamingRequest } from '../streamingRequest';
+import { RequestManager } from '../payloads';
 import {
     PayloadReceiver,
     PayloadSender
-} from '../PayloadTransport';
+} from '../payloadTransport';
 import { NamedPipeTransport } from './NamedPipeTransport';
-import { IStreamingTransportClient, IReceiveResponse } from '../Interfaces';
+import { IStreamingTransportClient, IReceiveResponse } from '../interfaces';
 
 export class NamedPipeClient implements IStreamingTransportClient {
     private readonly _baseName: string;
