+++ libraries/botframework-streaming/src/namedPipe/namedPipeTransport.ts
@@ -6,8 +6,8 @@
  * Licensed under the MIT License.
  */
 import { Socket } from 'net';
-import { ITransportSender } from '../Interfaces/ITransportSender';
-import { ITransportReceiver } from '../Interfaces/ITransportReceiver';
+import { ITransportSender } from '../interfaces/ITransportSender';
+import { ITransportReceiver } from '../interfaces/ITransportReceiver';
 
 export class NamedPipeTransport implements ITransportSender, ITransportReceiver {
     public static readonly PipePath: string = '\\\\.\\pipe\\';
