+++ libraries/botframework-streaming/src/payloadTransport/transportDisconnectedEventHandler.ts
@@ -5,6 +5,6 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-import { TransportDisconnectedEventArgs } from './TransportDisconnectedEventArgs';
+import { TransportDisconnectedEventArgs } from './transportDisconnectedEventArgs';
 
 export type TransportDisconnectedEventHandler = (sender: any, e: TransportDisconnectedEventArgs) => void;
