+++ libraries/botframework-streaming/src/payloads/headerSerializer.ts
@@ -5,8 +5,8 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-import { IHeader } from '../Interfaces/IHeader';
-import { PayloadConstants } from './PayloadConstants';
+import { IHeader } from '../interfaces/IHeader';
+import { PayloadConstants } from './payloadConstants';
 
 export class HeaderSerializer {
     public static readonly Delimiter = '.';
