+++ libraries/botframework-streaming/src/payloads/index.ts
@@ -6,10 +6,10 @@
  * Licensed under the MIT License.
  */
 
-export * from './HeaderSerializer';
-export * from './StreamManager';
-export * from './PayloadAssemblerManager';
-export * from './PayloadTypes';
-export * from './RequestManager';
-export * from './SendOperations';
-export * from './StreamManager';
+export * from './headerSerializer';
+export * from './streamManager';
+export * from './payloadAssemblerManager';
+export * from './payloadTypes';
+export * from './requestManager';
+export * from './sendOperations';
+export * from './streamManager';
