+++ libraries/botframework-streaming/src/subscribableStream.ts
