+++ libraries/botframework-streaming/src/webSocket/webSocketServer.ts
@@ -5,18 +5,18 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-import { ProtocolAdapter } from '../ProtocolAdapter';
-import { RequestHandler } from '../RequestHandler';
-import { StreamingRequest } from '../StreamingRequest';
-import { RequestManager } from '../Payloads';
+import { ProtocolAdapter } from '../protocolAdapter';
+import { RequestHandler } from '../requestHandler';
+import { StreamingRequest } from '../streamingRequest';
+import { RequestManager } from '../payloads';
 import {
     PayloadReceiver,
     PayloadSender,
     TransportDisconnectedEventArgs
-} from '../PayloadTransport';
-import { ISocket } from '../Interfaces/ISocket';
+} from '../payloadTransport';
+import { ISocket } from '../interfaces/ISocket';
 import { WebSocketTransport } from './WebSocketTransport';
-import { IStreamingTransportServer, IReceiveResponse } from '../Interfaces';
+import { IStreamingTransportServer, IReceiveResponse } from '../interfaces';
 
 /// <summary>
 /// A server for use with the Bot Framework Protocol V3 with Streaming Extensions and an underlying WebSocket transport.
