+++ libraries/botframework-streaming/src/webSocket/webSocketTransport.ts
@@ -5,9 +5,9 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-import { ISocket } from '../Interfaces';
-import { ITransportSender } from '../Interfaces/ITransportSender';
-import { ITransportReceiver } from '../Interfaces/ITransportReceiver';
+import { ISocket } from '../interfaces';
+import { ITransportSender } from '../interfaces/ITransportSender';
+import { ITransportReceiver } from '../interfaces/ITransportReceiver';
 
 export class WebSocketTransport implements ITransportSender, ITransportReceiver {
     private _socket: ISocket;
