+++ libraries/botframework-streaming/tests/ProtocolAdapter.test.js
@@ -1,14 +1,14 @@
-const PayloadAssembler = require('../lib/Assemblers/PayloadAssembler');
-const ProtocolAdapter = require('../lib/ProtocolAdapter');
-const RequestManager = require('../lib/Payloads/RequestManager');
-const PayloadSender = require('../lib/PayloadTransport/PayloadSender');
-const PayloadReceiver = require('../lib/PayloadTransport/PayloadReceiver');
-const PayloadConstants = require('../lib/Payloads/PayloadConstants');
-const SubscribableStream = require('../lib/SubscribableStream');
-const RequestHandler = require('../lib/RequestHandler');
-const Response = require('../lib/StreamingResponse');
-const Request = require('../lib/StreamingRequest');
-const StreamManager = require('../lib/Payloads/StreamManager');
+const PayloadAssembler = require('../lib/assemblers/payloadAssembler');
+const ProtocolAdapter = require('../lib/protocolAdapter');
+const RequestManager = require('../lib/payloads/requestManager');
+const PayloadSender = require('../lib/payloadTransport/payloadSender');
+const PayloadReceiver = require('../lib/payloadTransport/payloadReceiver');
+const PayloadConstants = require('../lib/payloads/payloadConstants');
+const SubscribableStream = require('../lib/subscribableStream');
+const RequestHandler = require('../lib/requestHandler');
+const Response = require('../lib/streamingResponse');
+const Request = require('../lib/streamingRequest');
+const StreamManager = require('../lib/payloads/streamManager');
 const  chai  = require('chai');
 var sinon = require('sinon');
 var expect = chai.expect;
