+++ libraries/botframework-streaming/tests/StreamManager.test.js
@@ -1,7 +1,7 @@
-const StreamManager = require('../lib/Payloads/StreamManager');
-const PayloadTypes = require('../lib/Payloads/PayloadTypes');
-const SubscribableStream = require('../lib/SubscribableStream');
-const PayloadAssembler = require('../lib/Assemblers/PayloadAssembler');
+const StreamManager = require('../lib/payloads/streamManager');
+const PayloadTypes = require('../lib/payloads/payloadTypes');
+const SubscribableStream = require('../lib/subscribableStream');
+const PayloadAssembler = require('../lib/assemblers/payloadAssembler');
 const  chai  = require('chai');
 var expect = chai.expect;
