+++ libraries/botframework-streaming/tests/SubscribableStream.test.js
@@ -1,4 +1,4 @@
-const Stream = require( '../lib/SubscribableStream');
+const Stream = require( '../lib/subscribableStream');
 const chai = require( 'chai');
 var expect = chai.expect;
